requirements.txt
openai-whisper==20230314
SpeechRecognition==3.10.0
PyAudio==0.2.13